Fix length parameter in send_feature_report
The length sent to hid_send_feature_report should always be the number of bytes in the buffer. send_feature_report is currently only sending the number of bytes after the report_id. From the hidapi documentation:

"Since the Report ID is mandatory, calls to hid_send_feature_report() will always contain one more byte than the report contains. For example, if a hid report is 16 bytes long, 17 bytes must be passed to hid_send_feature_report(): the Report ID (or 0x0, for devices which do not use numbered reports), followed by the report data (16 bytes). In this example, the length passed in would be 17."